app/requirements.txt
surprise==0.1
scikit-surprise==1.1.3
pandas==1.5.0
numpy==1.23.3
Flask==2.2.2
pandas-gbq==0.17.9
google-auth==2.12.0
google-auth-oauthlib==0.5.3
google-cloud-bigquery==3.3.3
tqdm==4.64.1
gunicorn==20.1.0
